Backend/requirements.txt
altair==5.2.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.14.0
bayesian-optimization==1.4.1
beautifulsoup4==4.12.3
bleach==6.1.0
blessings==1.7
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.5
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==2.2.0
dj-rest-auth==5.1.0
Django>=3.2,<5.0
django-allauth==0.61.1
django-bootstrap5==23.4
django-cors-headers==3.2.1
django-mathfilters==1.0.0
django-tinymce==3.6.1
djangorestframework==3.15.1
djangorestframework-simplejwt==4.7.1
docopt==0.6.2
drf-yasg==1.21.4
ee==0.2
et-xmlfile==1.1.0
executing==2.0.1
fastjsonschema==2.19.1
fonttools==4.47.0
fqdn==1.5.1
gitdb==4.0.11
GitPython==3.1.40
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
importlib-metadata==6.11.0
inflection==0.5.1
ipykernel==6.29.3
ipython==8.22.2
ipywidgets==8.1.2
isoduration==20.11.0
itypes==1.2.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.4.2
json5==0.9.22
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.4
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.13.0
jupyter_server_terminals==0.5.2
jupyterlab==4.1.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.3
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mysqlclient==2.2.4
nbclient==0.9.0
nbconvert==7.16.2
nbformat==5.9.2
nest-asyncio==1.6.0
notebook==7.1.1
notebook_shim==0.2.4
numpy==1.26.2
oauthlib==3.2.2
openpyxl==3.1.2
overrides==7.7.0
packaging==23.2
pandas==2.1.4
pandocfilters==1.5.1
parso==0.8.3
pillow==10.3.0
pipreqs==0.4.13
pivottablejs==0.9.0
platformdirs==4.2.0
plotly==5.18.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.1
psutil==5.9.8
psycopg2==2.9.9
pure-eval==0.2.2
pyarrow==14.0.2
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
PyMySQL==1.1.1
pyparsing==3.1.1
pypdf==4.2.0
PyPDF2==3.0.1
pypng==0.20220715.0
python-dateutil==2.8.2
python-docx==1.1.0
python-json-logger==2.0.7
python3-openid==3.2.0
pytz==2023.3.post1
pywin32==306; sys_platform == 'win32'
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==25.1.2
qrcode==7.4.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.32.0
reportlab==4.1.0
requests==2.31.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.15.2
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
Send2Trash==1.8.2
setuptools==69.5.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.5.0
stack-data==0.6.3
streamlit==1.29.0
tabulate==0.9.0
tenacity==8.2.3
terminado==0.18.0
threadpoolctl==3.5.0
tika==2.6.0
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.0
tornado==6.4
traitlets==5.14.1
types-python-dateutil==2.8.19.20240311
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
widgetsnbextension==4.0.10
wordcloud==1.9.3
yarg==0.1.9
zipp==3.17.0



